requirements.txt
-e .
APScheduler==3.7.0
Brotli==1.0.9
Click==7.0
Flask-APScheduler==1.12.1
Flask-Compress==1.9.0
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
Flask-SSLify==0.1.5
Flask==1.1.1
Jinja2==2.11.0
MarkupSafe==1.1.1
PyYAML==5.3
SQLAlchemy==1.3.22
Werkzeug==0.16.1
aiohttp==3.7.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
azure-core==1.10.0
azure-eventhub-checkpointstoreblob-aio==1.1.2
azure-eventhub==5.3.0
azure-storage-blob==12.7.1
bidict==0.21.2
blis==0.4.1
boto3==1.17.22
botocore==1.20.22
cachetools==4.2.1
catalogue==0.0.8
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
colorhash==1.0.3
configparser==5.0.1
cryptography==3.4.1
cycler==0.10.0
cymem==2.0.3
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
dill==0.3.3
dnspython==1.16.0
future==0.18.2
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
gunicorn==20.0.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
idna-ssl==1.1.0
idna==2.10
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==1.4.0
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==1.1.0
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
matplotlib==3.1.2
more-itertools==8.1.0
msrest==0.6.21
multidict==5.1.0
murmurhash==1.0.2
nest-asyncio==1.5.1
numpy==1.18.1
oauthlib==3.1.0
packaging==20.9
pandas==0.25.3
passlib==1.7.4
plac==1.1.3
plotly==4.14.3
pluggy==0.13.1
preshed==3.0.2
protobuf==3.15.6
psutil==5.8.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pymongo==3.11.0
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==6.2.2
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2019.3
regexp==0.1
requests-oauthlib==1.3.0
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
s3transfer==0.3.4
scikit-learn==0.22.1
scipy==1.5.4
semantic-version==2.8.5
setuptools-rust==0.11.6
six==1.14.0
sklearn-pandas==1.8.0
sklearn==0.0
spacy==2.2.3
srsly==0.2.0
thinc==7.3.1
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
uamqp==1.2.14
urllib3==1.26.3
wasabi==0.4.2
wincertstore==0.2
xgboost==0.90
yarl==1.6.3
zipfile36==0.1.3
zipp==2.0.1
zope.event==4.5.0
zope.interface==5.3.0
dvc
dvc[gdrive]
mlflow 
astrapy==0.0.2
cassandra-driver==3.25.0
